add more update to style
